Amended inotify script and update_song.py command to handle files inside dirs when added and removed. Remains to add directory view to Tracks Tree.

+++ musiclib/management/commands/update_song.py
@@ -11,21 +11,37 @@ import mutagen
 This command is to be called by the inotify script whenever a song is added or deleted in the audio dir. It updates the database accordingly. 
 '''
 class Command(BaseCommand):
-    def add_song(self, filename, seconds):
+    def add_song(self, fullpath, filename, seconds):
         # TODO add code to preprocess song file for metadata (song name, artist, album, genre maybe?)
         #with open(os.path.join(settings.MEDIA_ROOT, filename), 'rb') as f:
             #songf = File(f, name=filename)
 
-        song_metadata = mutagen.File(os.path.join(settings.MEDIA_ROOT, filename))
+        root_path = settings.MEDIA_ROOT
 
-        new_song = Song(song_file=filename, time_added=seconds)
+        song_fullname = os.path.join(fullpath, filename)
+        song_relativename = song_fullname[len(root_path):].lstrip('/')
+        print("Relative name: ", song_relativename)
+        song_metadata = mutagen.File(song_fullname)
         
-       
-        new_song.name = song_metadata.get('title', [None])[0]
 
-        date = song_metadata.get('date', [None])[0] 
+        all_songs = Song.objects.all()
+        new_song = None
+        for song in all_songs:
+            if song.song_file.path == song_fullname:
+                # if song already exists, update its metadata instead of creating new entry in DB
+                new_song = song
+                break
 
-        artist_name = song_metadata.get('artist', [None])[0]
+        if not new_song:
+            new_song = Song(song_file=song_relativename, time_added=seconds)
+        
+        if song_metadata:
+            new_song.name = song_metadata.get('title', [None])[0]
+            date = song_metadata.get('date', [None])[0] 
+            artist_name = song_metadata.get('artist', [None])[0]
+            album_name = song_metadata.get('album', [None])[0]
+        else:
+            new_song.name = date = artist_name = album_name = None
 
         if artist_name:
             try:
@@ -36,7 +52,7 @@ class Command(BaseCommand):
 
             new_song.artist = artist
 
-        album_name = song_metadata.get('album')
+
 
         if album_name:
             try:
@@ -52,24 +68,52 @@ class Command(BaseCommand):
 
     def remove_song(self, filename):
         song_to_delete = Song.objects.get(song_file=filename)
-        song_to_delete.delete()
+        if song_to_delete:
+            song_to_delete.delete()
 
+    def remove_songs_rooted_at(self, fulldirname):
+        all_songs = Song.objects.all()
+
+        for song in all_songs:
+            if song.song_file.path.startswith(fulldirname):
+               song.delete()
 
     def add_arguments(self, parser):
         parser.add_argument('event')
         parser.add_argument('seconds', type=int)
         parser.add_argument('filename')
+        parser.add_argument('fullpath')
 
     def handle(self, *args, **options): 
-        event = options['event'] 
+        events = options['event'].split(",")
+        if len(events) > 1:
+            event, isdir = events
+        else:
+            isdir = None
+            event = events[0]
+
         seconds = options['seconds'] 
-        filename = options['filename'] 
+        filename = options['filename']
+        fullpath = options['fullpath']
 
-        print("From update_song.py:", event, seconds, filename, sep=' ')
+        print("From update_song.py:", event, isdir, seconds, fullpath, filename, sep='\n', end='\n\n')
 
-        if((event == 'MOVED_FROM') | (event == 'DELETE')):
+
+        if isdir == 'ISDIR':
+            fulldirname = os.path.join(fullpath, filename)
+            if(event == 'MOVED_FROM' or event == 'DELETE'):
+                self.remove_songs_rooted_at(fulldirname)
+                
+            elif(event == 'CREATE' or event == 'MOVED_TO'):
+                for fullpath, dirs, filenames in os.walk(fulldirname):
+                    for fname in filenames:
+                        print("Full path: ", fulldirname, "Filename: ", fname, end="\n\n")
+                        # TODO check if it's actually an audio file
+                        self.add_song(fulldirname, fname, seconds)
+
+        elif((event == 'MOVED_FROM') or (event == 'DELETE')):
             self.remove_song(filename)
-        elif((event == 'CREATE') | (event == 'MOVED_TO')):
-            self.add_song(filename, seconds)
+        elif((event == 'CREATE') or (event == 'MOVED_TO')):
+            self.add_song(fullpath, filename, seconds)
